Move two files to the Integrations project
